+++ src/aliases/utility/toggle_pause.lua
@@ -1,1 +1,0 @@
-mapper.pause(matches[2])